requirements.txt
Django==3.1.3
djangorestframework==3.12.2
Pillow==8.0.1
PyYAML==5.3.1
uritemplate==3.0.1
